@@ -2,6 +2,7 @@
   "name": "@borkenware/spoonfeed",
   "version": "0.1.0",
   "description": "Generate beautiful documentation portals in seconds from easy-to-edit markdown files.",
+  "type": "module",
   "main": "dist/spoonfeed.js",
   "repository": "git@github.com:borkenware/spoonfeed.git",
   "author": "Cynthia <cyyynthia@borkenware.com>",
@@ -13,45 +14,32 @@
     "spoonfeed": "dist/cli/index.js"
   },
   "scripts": {
-    "test": "jest",
-    "lint": "eslint src postbuild.js",
+    "lint": "eslint src --ext ts",
     "build": "tsc && node postbuild.js"
   },
   "dependencies": {
-    "@babel/core": "^7.12.7",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-runtime": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
     "@babel/runtime": "^7.12.5",
-    "@rollup/plugin-babel": "^5.2.1",
-    "@rollup/plugin-commonjs": "^16.0.0",
-    "@rollup/plugin-node-resolve": "^10.0.0",
-    "preact": "^10.5.7",
+    "@rollup/plugin-babel": "^5.2.2",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.1.0",
+    "preact": "^10.5.11",
     "preact-helmet": "^4.0.0-alpha-3",
     "preact-router": "^3.2.1",
-    "rollup": "^2.33.3",
+    "rollup": "^2.38.0",
     "rollup-plugin-terser": "^7.0.2",
-    "sass": "^1.29.0",
+    "sass": "^1.32.5",
     "shiki": "^0.2.7"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.9",
+    "@types/node": "^14.14.22",
     "@types/sass": "^1.16.0",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
-    "babel-jest": "^26.6.3",
-    "eslint": "^7.14.0",
-    "eslint-import-resolver-typescript": "^2.3.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-optimize-regex": "^1.2.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "^26.6.3",
-    "memfs": "^3.2.0",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.1.2"
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.1",
+    "eslint": "^7.18.0",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   }
 }
